+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/venue/Venue.kt
@@ -1,0 +1,20 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.venue
+
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CommonVenueData
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.Locationed
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import kotlinx.serialization.*
+
+@Serializable
+data class Venue(
+    @SerialName(locationField)
+    val location: Location,
+    @SerialName(titleField)
+    override val title: String,
+    @SerialName(addressField)
+    override val address: String,
+    @SerialName(foursquareIdField)
+    override val foursquareId: FoursquareId? = null,
+    @SerialName(foursquareTypeField)
+    override val foursquareType: FoursquareType? = null
+) : CommonVenueData, Locationed by location